requirements.txt
py-cord[voice]
youtube_dl
aio-pika
python-dotenv